@@ -1,68 +1,86 @@
 {
-  "name": "forwarder-ui",
-  "version": "1.0.21",
+  "name": "data-forwarders",
+  "version": "1.0.22",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "webpack-dev-server",
-    "build": "webpack",
-    "deploy": "npm run build",
-    "test": "mocha --compilers js:babel/register --recursive",
-    "test:watch": "npm test -- --watch"
+    "clean": "rimraf dist",
+    "build:webpack": "NODE_ENV=production webpack --config webpack.config.prod.js",
+    "build": "npm run clean && npm run build:webpack",
+    "lint": "eslint src",
+    "mocha": "NODE_ENV=test mocha",
+    "start": "node dev-server.js",
+    "test": "npm run mocha",
+    "test:watch": "npm run mocha -- -w -R mocha-multi --reporter-options spec=-,mocha-osx-reporter=-"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/octoblu/forwarder-ui.git"
+    "url": "https://github.com/octoblu/data-forwarders.git"
   },
-  "author": "",
+  "author": "Octoblu, Inc",
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/octoblu/forwarder-ui/issues"
+    "url": "https://github.com/octoblu/data-forwarders/issues"
   },
-  "homepage": "https://github.com/octoblu/forwarder-ui",
-  "devDependencies": {
-    "autoprefixer-core": "^6.0.1",
-    "babel": "^5.8.23",
-    "babel-core": "^5.8.25",
-    "babel-loader": "^5.3.2",
-    "coffee-loader": "^0.7.2",
-    "coffee-script": "^1.10.0",
-    "css-loader": "^0.18.0",
-    "expect": "^1.10.0",
-    "file-loader": "^0.8.4",
-    "hjs-webpack": "^2.13.1",
-    "json-loader": "^0.5.3",
-    "mocha": "^2.3.3",
-    "mocha-osx-reporter": "^0.1.2",
-    "postcss-loader": "^0.6.0",
-    "react": "^0.13.3",
-    "react-hot-loader": "^1.3.0",
-    "redux-devtools": "^2.1.2",
-    "style-loader": "^0.12.4",
-    "stylus-loader": "^1.3.0",
-    "url-loader": "^0.5.6",
-    "webpack": "^1.12.2",
-    "webpack-dev-server": "^1.11.0",
-    "yeticss": "^7.0.5"
-  },
+  "homepage": "https://github.com/octoblu/data-forwarders",
   "dependencies": {
-    "autoprefixer": "^6.0.3",
-    "classnames": "^2.1.3",
-    "color": "^0.10.1",
-    "css-loader": "^0.18.0",
-    "history": "^1.9.1",
-    "lodash": "^3.10.1",
-    "meshblu": "^1.30.1",
-    "meshblu-http": "^3.4.1",
-    "postcss-loader": "^0.6.0",
-    "radium": "^0.14.1",
-    "react-redux": "^2.1.2",
-    "react-router": "^1.0.0-rc1",
-    "redux": "^3.0.0",
-    "redux-devtools": "^2.1.2",
-    "redux-react-router": "^1.0.0-beta3",
-    "redux-thunk": "^1.0.0",
-    "style-loader": "^0.12.4",
-    "yeticss": "^7.0.5"
+    "atob": "^2.0.3",
+    "isomorphic-fetch": "^2.2.1",
+    "meshblu": "^1.32.0",
+    "react": "^15.0.1",
+    "react-cookie": "^0.4.5",
+    "react-dom": "^15.0.1",
+    "react-redux": "^4.4.0",
+    "react-router": "^2.0.0",
+    "react-router-redux": "^4.0.4",
+    "react-tabs": "^0.5.3",
+    "redux": "^3.5.2",
+    "redux-thunk": "^2.0.1",
+    "zooid-meshblu-device-editor": "^5.0.0",
+    "zooid-ui": "^4.0.1"
+  },
+  "devDependencies": {
+    "@octoblu/shmock": "^0.7.2",
+    "autoprefixer": "^6.3.3",
+    "babel-core": "^6.6.5",
+    "babel-eslint": "^6.0.4",
+    "babel-loader": "^6.2.4",
+    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-react": "^6.5.0",
+    "babel-preset-react-hmre": "^1.1.1",
+    "babel-preset-stage-0": "^6.5.0",
+    "babel-register": "^6.7.2",
+    "chai": "^3.5.0",
+    "chai-enzyme": "^0.4.2",
+    "cross-env": "^1.0.7",
+    "css-loader": "^0.23.1",
+    "enzyme": "^2.2.0",
+    "eslint": "^2.8.0",
+    "eslint-config-airbnb": "^8.0.0",
+    "eslint-plugin-babel": "^3.2.0",
+    "eslint-plugin-import": "^1.6.0",
+    "eslint-plugin-jsx-a11y": "^1.0.3",
+    "eslint-plugin-react": "^5.0.1",
+    "eventsource-polyfill": "^0.9.6",
+    "express": "^4.13.4",
+    "jsdom": "^8.4.0",
+    "json-loader": "^0.5.4",
+    "mocha": "^2.4.5",
+    "mocha-multi": "^0.9.0",
+    "mocha-osx-reporter": "^0.1.2",
+    "nock": "^8.0.0",
+    "postcss-loader": "^0.8.1",
+    "postcss-scss": "^0.1.7",
+    "precss": "^1.4.0",
+    "react-addons-test-utils": "^15.0.1",
+    "redux-logger": "^2.6.1",
+    "redux-mock-store": "^1.0.2",
+    "rimraf": "^2.5.2",
+    "sinon": "^1.17.3",
+    "sinon-chai": "^2.8.0",
+    "style-loader": "^0.13.0",
+    "webpack": "^1.12.14",
+    "webpack-dev-middleware": "^1.5.1",
+    "webpack-hot-middleware": "^2.9.1"
   }
 }
