@@ -1,25 +1,15 @@
 import React from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
+import { destroyAuthentication } from '../services/auth-service'
 
-// Actions
-import * as MeshbluActions from '../actions/meshblu-actions';
-
-var Logout = React.createClass({
-  componentWillMount: function() {
-    const { dispatch, meshblu } = this.props;
-    dispatch(MeshbluActions.closeConnection(meshblu.connection));
-  },
-
-  render: function(){
-    return(<div></div>);
+export default class Logout extends React.Component {
+  componentWillMount() {
+    destroyAuthentication();
+    window.location = '/';
   }
-});
 
-function mapStateToProps(state) {
-  return {
-    meshblu: state.meshblu
-  };
+  render(){
+    return null;
+  }
 }
 
-export default connect(mapStateToProps)(Logout);
+export default Logout
