@@ -1,40 +1,40 @@
-export const FETCH_DEVICES_REQUEST                  = 'FETCH_DEVICES_REQUEST';
-export const FETCH_DEVICES_SUCCESS                  = 'FETCH_DEVICES_SUCCESS';
-export const FETCH_DEVICES_ERROR                    = 'FETCH_DEVICES_ERROR';
+export const FETCH_FORWARDERS_REQUEST = 'FETCH_FORWARDERS_REQUEST';
+export const FETCH_FORWARDERS_SUCCESS = 'FETCH_FORWARDERS_SUCCESS';
+export const FETCH_FORWARDERS_FAILURE = 'FETCH_FORWARDERS_FAILURE';
 
-export const FETCH_DEVICE_DETAIL                    = 'FETCH_DEVICES_DETAIL';
+export const FETCH_FORWARDER_BY_UUID_REQUEST = 'FETCH_FORWARDER_BY_UUID_REQUEST'
+export const FETCH_FORWARDER_BY_UUID_SUCCESS = 'FETCH_FORWARDER_BY_UUID_SUCCESS'
+export const FETCH_FORWARDER_BY_UUID_FAILURE = 'FETCH_FORWARDER_BY_UUID_FAILURE'
 
-export const MESHBLU_CREATE_CONNECTION              = 'MESHBLU_CREATE_CONNECTION';
-export const MESHBLU_CREATE_CONNECTION_REQUEST      = 'MESHBLU_CREATE_CONNECTION_REQUEST';
-export const MESHBLU_CREATE_CONNECTION_SUCCESS      = 'MESHBLU_CREATE_CONNECTION_SUCCESS';
-export const MESHBLU_CREATE_CONNECTION_ERROR        = 'MESHBLU_CREATE_CONNECTION_ERROR';
+export const DELETE_FORWARDER_BY_UUID_REQUEST = 'DELETE_FORWARDER_BY_UUID_REQUEST'
+export const DELETE_FORWARDER_BY_UUID_SUCCESS = 'DELETE_FORWARDER_BY_UUID_SUCCESS'
+export const DELETE_FORWARDER_BY_UUID_FAILURE = 'DELETE_FORWARDER_BY_UUID_FAILURE'
 
-export const MESHBLU_CLOSE_CONNECTION               = 'MESHBLU_CLOSE_CONNECTION';
-export const MESHBLU_REMOVE_CONNECTION              = 'MESHBLU_REMOVE_CONNECTION';
+export const CREATE_FORWARDER_REQUEST = 'CREATE_FORWARDER_REQUEST';
+export const CREATE_FORWARDER_SUCCESS = 'CREATE_FORWARDER_SUCCESS';
+export const CREATE_FORWARDER_FAILURE = 'CREATE_FORWARDER_FAILURE';
 
-export const MESHBLU_REGISTER_DEVICE_REQUEST        = 'MESHBLU_REGISTER_DEVICE_REQUEST';
-export const MESHBLU_REGISTER_DEVICE_SUCCESS        = 'MESHBLU_REGISTER_DEVICE_SUCCESS';
-export const MESHBLU_REGISTER_DEVICE_ERROR          = 'MESHBLU_REGISTER_DEVICE_ERROR';
+export const FETCH_TYPES_REQUEST = 'FETCH_TYPES_REQUEST';
+export const FETCH_TYPES_SUCCESS = 'FETCH_TYPES_SUCCESS';
+export const FETCH_TYPES_FAILURE = 'FETCH_TYPES_FAILURE';
 
-export const MESHBLU_DELETE_DEVICE_REQUEST          = 'MESHBLU_DELETE_DEVICE_REQUEST';
-export const MESHBLU_DELETE_DEVICE_SUCCESS          = 'MESHBLU_DELETE_DEVICE_SUCCESS';
-export const MESHBLU_DELETE_DEVICE_ERROR            = 'MESHBLU_DELETE_DEVICE_ERROR';
+export const FETCH_MY_DEVICES_REQUEST = 'FETCH_MY_DEVICES_REQUEST';
+export const FETCH_MY_DEVICES_SUCCESS = 'FETCH_MY_DEVICES_SUCCESS';
+export const FETCH_MY_DEVICES_FAILURE = 'FETCH_MY_DEVICES_FAILURE';
 
-export const MESHBLU_UPDATE_DEVICE_REQUEST          = 'MESHBLU_UPDATE_DEVICE_REQUEST';
-export const MESHBLU_UPDATE_DEVICE_SUCCESS          = 'MESHBLU_UPDATE_DEVICE_SUCCESS';
-export const MESHBLU_UPDATE_DEVICE_ERROR            = 'MESHBLU_UPDATE_DEVICE_ERROR';
 
-export const FORWARDER_ADD_DATA_STORE               = 'FORWARDER_ADD_DATA_STORE';
-export const FORWARDER_SET_NAME                     = 'FORWARDER_SET_NAME';
-export const FORWARDER_SET_OWNER                    = 'FORWARDER_SET_OWNER';
-export const FORWARDER_SET_OPTIONS                  = 'FORWARDER_SET_OPTIONS';
-export const FORWARDER_SET_OPTIONS_CANCELED         = 'FORWARDER_SET_OPTIONS_CANCELED';
-export const FORWARDER_SET_OPTIONS_VALUE            = 'FORWARDER_SET_OPTIONS_VALUE';
-export const FORWARDER_SET_OPTIONS_ERROR            = 'FORWARDER_SET_OPTIONS_ERROR';
-export const FORWARDER_SHOW_DETAIL_ERROR            = 'FORWARDER_SHOW_DETAIL_ERROR';
-export const FORWARDER_ADD_GATEBLU                  = 'FORWARDER_ADD_GATEBLU';
-export const FORWARDER_SUBSCRIBE_TO_DEVICE          = 'FORWARDER_SUBSCRIBE_TO_DEVICE';
-export const FORWARDER_SUBSCRIBE_TO_ALL_DEVICES     = 'FORWARDER_SUBSCRIBE_TO_ALL_DEVICES';
-export const FORWARDER_UNSUBSCRIBE_FROM_DEVICE      = 'FORWARDER_UNSUBSCRIBE_FROM_DEVICE';
-export const FORWARDER_UNSUBSCRIBE_FROM_ALL_DEVICES = 'FORWARDER_UNSUBSCRIBE_FROM_ALL_DEVICES';
-export const FORWARDER_INITIALIZE                   = 'FORWARDER_INITIALIZE';
+export const CREATE_SUBSCRIPTION_REQUEST = 'CREATE_SUBSCRIPTION_REQUEST';
+export const CREATE_SUBSCRIPTION_SUCCESS = 'CREATE_SUBSCRIPTION_SUCCESS';
+export const CREATE_SUBSCRIPTION_FAILURE = 'CREATE_SUBSCRIPTION_FAILURE';
+
+
+export const DELETE_SUBSCRIPTION_REQUEST = 'DELETE_SUBSCRIPTION_REQUEST';
+export const DELETE_SUBSCRIPTION_SUCCESS = 'DELETE_SUBSCRIPTION_SUCCESS';
+export const DELETE_SUBSCRIPTION_FAILURE = 'DELETE_SUBSCRIPTION_FAILURE';
+
+export const FETCH_FORWARDER_TYPE_CONFIG_REQUEST = 'FETCH_FORWARDER_TYPE_CONFIG_REQUEST';
+export const FETCH_FORWARDER_TYPE_CONFIG_SUCCESS = 'FETCH_FORWARDER_TYPE_CONFIG_SUCCESS';
+export const FETCH_FORWARDER_TYPE_CONFIG_FAILURE = 'FETCH_FORWARDER_TYPE_CONFIG_FAILURE';
+
+export const SET_ACTIVE_FORWARDER_TYPE   = 'SET_ACTIVE_FORWARDER_TYPE';
+export const UNSET_ACTIVE_FORWARDER_TYPE = 'UNSET_ACTIVE_FORWARDER_TYPE';
