@@ -1,16 +1,16 @@
-import { combineReducers } from 'redux';
-import devices from './devices-reducers';
-import dataStores from './data-stores-reducers';
-import forwarder from './forwarder-reducers';
-import meshblu from './meshblu-reducers';
-import { routerStateReducer } from 'redux-react-router';
+import { combineReducers } from 'redux'
+import { routerReducer } from 'react-router-redux'
+import forwardersReducer from './forwarders/forwarder-reducer';
+import typesReducer from './types/types-reducer';
+import activeForwarderTypeReducer from './active-type/active-type-reducer';
+import myDevicesReducer from './my-devices/my-devices-reducer';
 
 const rootReducer = combineReducers({
-  devices,
-  dataStores,
-  forwarder,
-  meshblu,
-  router: routerStateReducer
-});
+  activeForwarderType: activeForwarderTypeReducer,
+  forwarders: forwardersReducer,
+  myDevices: myDevicesReducer,
+  routing: routerReducer,
+  types: typesReducer,
+})
 
-export default rootReducer;
+export default rootReducer
